Core Files

+++ autoupdater/scripts/startup.js
@@ -3,11 +3,13 @@ global.sharedObj = {version: null, startMinecraft: null, resetConfig: null, redo
 global.sharedObj.version = "0.1.2r";
 global.sharedObj.resetConfig = function () {
 	deleteFolderRecursive("./minecraft/.minecraft/config/Embassy");
-	fs.unlinkSync("./minecraft/.minecraft/config/betterhud.cfg");
-	console.log("Pulling config");
-	downloadFile('https://remethor.github.io/autoupdater/config/betterhud.cfg', './minecraft/.minecraft/config/betterhud.cfg', function(){
+	if(fs.existsSync("./minecraft/.minecraft/config/betterhud.cfg")){
+		fs.unlinkSync("./minecraft/.minecraft/config/betterhud.cfg");
+		console.log("Pulling config");
+		downloadFile('https://remethor.github.io/autoupdater/config/betterhud.cfg', './minecraft/.minecraft/config/betterhud.cfg', function(){
 		console.log("Done!");
 	});
+	}
 }
 global.sharedObj.resetConfig();
 global.sharedObj.startMinecraft = function () {
